refactor: extract WatermarkSettingsForm from the download modal
